Fix build: move non-async exports out of 'use server' module
A 'use server' file can only export async functions; the DISCOVERABLE_VISIBILITIES
const and PlayerEventOverviewItem type broke page-data collection for every route
importing event-artifact.ts (e.g. /api/events/[id]/calendar). Moved them to
src/lib/event-discoverability.ts.
